requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2
django-apscheduler==0.6.2
django-auth-ldap==4.6.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
exceptiongroup==1.2.0
idna==3.4
iniconfig==2.0.0
ldap3==2.9.1
MarkupSafe==2.1.2
numpy==1.24.2
packaging==23.1
Pillow==9.5.0
pluggy==1.2.0
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycountry==22.3.5
pycparser==2.21
PyGithub==1.58.1
PyJWT==2.6.0
PyNaCl==1.5.0
pytest==7.4.0
python-dateutil==2.8.2
python-gitlab==3.14.0
python-ldap==3.4.4
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.12
requests==2.30.0
requests-toolbelt==1.0.0
schedule==1.2.0
six==1.16.0
sqlparse==0.4.3
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.2
wrapt==1.15.0
